requirements.txt
streamlit
pandas
python-dotenv
pymysql
requests
cryptography
Pillow
Werkzeug
